MP Simple extension: control message
Former system suffered from the problem that they may start not
synchronous. It can then happen that messages get lost during boot
when other tiles are not ready.

With this patch a sender now has to check for the availability of the
receiver before sending messages. Therefore it sends a single flit
packet of the format:

 +---+--------+-----+----------+---+---+
 | S | SENDER | 111 | RECEIVER | * | 0 |
 +---+--------+-----+----------+---+---+

The network adapter replies with a single flit packet:

 +---+--------+-----+----------+---+---+---+
 | S | SENDER | 111 | RECEIVER | * | R | 1 |
 +---+--------+-----+----------+---+---+---+

where R signals whether the NA is ready (R=1 is ready).

To activate the device the software writes to the offset 0x4.
Afterwards all control requests are answered with R=1.

It is necessary to activate the sender device before sending control
messages to other tiles.